Fixed Bluetooth 4 frameTypeOffset location
After moving from a hardcoded frameTypeOffset to a  dynamic one to support cross-platform (lin/win) captures, it broke BT4 offset.  This is a change to look for the "signature" (frameTypeOffset 0, length 56) for a BT frame.  It's a bit of a hack, but it works for now -- I'm open to alternative ideas.